// server/src/common/enums/ModelType.enum.ts
export enum ModelType {
  GPT_4O = 'gpt-4o',
  GPT_4_1 = 'gpt-4.1',
  GPT_4O_MINI = 'gpt-4o-mini',
  O3_MINI = 'o3-mini',
  O4_MINI = 'o4-mini',
  DEEPSEEK_CHAT = 'deepseek-chat',
  DEEPSEEK_REASONER = 'deepseek-reasoner',
  GEMINI_2_0_FLASH = 'gemini-2.0-flash',
  GEMINI_2_5_PRO = 'gemini-2.5-pro-preview-03-25',
}